podresources: add endpoint to query all devices
To enable topology aware scheduling, the reporting agent wants
to learn about the available devices on the worker node, in
order to then let the scheduler know and do an informed decision.

As it is now, the kubelet is the only authoritative source for
the resource accounting, for example because it manages the devices
plugin. Thus we need to exfiltrate this information from the kubelet.

A good fit for this task is the podresources API.
We add a new endpoint to let clients enumerate the resources.

Signed-off-by: Francesco Romani <fromani@redhat.com>