podresources: watch: fix WatchPodResourcesResponse
Per comments during the review of the Watch endpoint proposal:
1. https://github.com/kubernetes/enhancements/pull/1926#discussion_r466735964
2. https://github.com/kubernetes/enhancements/pull/1926#pullrequestreview-498559155

The agreed semantic of a Watch() response message
a. refers to a single pod. The intention was always to stream only individual pod changes
   (https://github.com/kubernetes/enhancements/pull/1926#discussion_r467081551)
b. must allow the client to reconcile with the response of the List()
   endpoint, thus must include a pod resource version.

This patch thus adds the missing resource version field and removes the
`repeated` attribute to the `PodResources` field.
Removing `repeated` is the simplest possible change that aligns the
proposal to the intention.

Alternatively, it is possible to change the proto so we can allow a
`WarchPodResourcesResponse` object to convey information about more
pods; however the performance and UX benefits of this more invasive
change are unclear, so we avoid it at this moment.

This change was missing because it was lost in a rebase

Signed-off-by: Francesco Romani <fromani@redhat.com>