README: downsize the kep
We need to postpone the Watch() part because it needs more design
work and for capacity reasons.
The KEP reflects now the current state of work, and the minimal
changes reflected in https://github.com/kubernetes/kubernetes/pull/95734

The GetAllocatableResources API is the minimum needed to enable
topology aware scheduling.

Signed-off-by: Francesco Romani <fromani@redhat.com>