Update implementation details in the KEP
Based on the review comments received in the implementaion PR
https://github.com/kubernetes/kubernetes/pull/101432, the KEP
is updated to reflect the most upto date implementation strategy.

This patch also captures suggestions provided during the first
round of review of the KEP and transitions from use of `smtaware`
as CPU Manager policy value to the use of another kubelet flag called
`cpumanager-policy-options` which when specified with `smtalign`
(instead of smtaware) as the value, enables this feature performs
the check in the static CPU Manager policy allocation logic to strictly
guranteed allocation of a full core.

Update from the term smtaware to smtalign throughout the KEP.

Co-authored-by: Francesco Romani <fromani@redhat.com>
Signed-off-by: Swati Sehgal <swsehgal@redhat.com>